requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
pandas>=2.1.0
tqdm>=4.66.0
pymysql>=1.1.0
python-dotenv>=1.0.0 